Refactor logging and server management components for improved clarity and functionality
- Removed the timestamp from the LogEntry initialization in ServerManager to streamline log entry creation.
- Enhanced the LogEntry structure by adding a category property for better log categorization.
- Updated LogLevel to include a display name for improved readability in the UI.
- Cleaned up the ServerManagerProcess by removing unused methods related to process output handling and environment setup, enhancing maintainability.

These changes aim to improve the logging system and server management processes, making them more efficient and user-friendly.
